move DatasetError exception classes to own file
