+++ api/scpca_portal/exceptions.py
@@ -3,28 +3,6 @@ from typing import Iterable
 from scpca_portal.enums import Modalities
 
 
-class DatasetError(Exception):
-    def __init__(self, message: str | None = None, dataset=None):
-        default_message = "A dataset error occurred."
-
-        message = message or default_message
-        if dataset:
-            message = f"Dataset {dataset.id}: {message}"
-        super().__init__(message)
-
-
-class DatasetLockedProjectError(DatasetError):
-    def __init__(self, dataset=None):
-        message = "Dataset has a locked project."
-        super().__init__(message, dataset)
-
-
-class DatasetMissingLibrariesError(DatasetError):
-    def __init__(self, dataset=None):
-        message = "Unable to find libraries for Dataset."
-        super().__init__(message, dataset)
-
-
 class JobError(Exception):
     def __init__(self, message: str | None = None, job=None):
         default_message = "A job error occurred."
