+++ api/scpca_portal/exceptions/__init__.py
@@ -1,0 +1,5 @@
+from scpca_portal.exceptions.dataset_error import (
+    DatasetError,
+    DatasetLockedProjectError,
+    DatasetMissingLibrariesError,
+)
